fix(inverted): 倒排/位图 64 位化（P79）——多表高位 docid 全链 u64
§26 多表 docid=table_id<<48|row 后非默认表 docid>=2^48，倒排 posting/位图全链按
u32（RoaringBitmap）实现：写路径断言崩溃、读路径 as u32 截断致检索错误。
- inverted.rs 统一 RoaringTreemap（type Posting）；SEG_VERSION 5->6（treemap 序列化，
  v2-v5 旧段按版本解码 32 位后 bm32() 升 64 位，默认表零迁移）；search/doc_count/
  search_paged(Vec<u64>)/iter_terms/位图 AND/GC/游标/merge_distinct 全升位
- engine.inverted_posting -> RoaringTreemap；sqlish 别名整域换 treemap；rpc docid
  Vec<u64>；gateway/shard_inverted 兼容适配
验证：lib cargo test 628/0 过；端到端非默认表 t + status/region 位图装载/37 探针全过；
遗留：多表单删须关 deletion_bitmap（§26 M3 已记载约束）

+++ src/rpc.rs
@@ -309,7 +309,7 @@ pub fn register_shard_handlers(server: &RpcServer, engine: Arc<Mutex<Engine>>) {
                 .ok_or_else(|| "shard.search_docids 缺少 term".to_string())?;
             let mut eng = engine.lock().unwrap();
             let posting = eng.inverted_posting(term).map_err(|e| e.to_string())?;
-            let docids: Vec<u32> = posting.iter().collect();
+            let docids: Vec<u64> = posting.iter().collect();
             Ok(json!({"docids": docids}))
         }
     });
